Case Study

+++ insight-server/app/controller/deliveryAddress.js
@@ -1,13 +1,15 @@
-const DeliveryAddress = require("../model/deliveryAddress");
+const { subject } = require("@casl/ability");
+const DeliveryAddress = require("../model/deliveryAddressModel");
+const { policyFor } = require("../../utils/policies");
 
 const getDeliveryAddress = async (req, res, next) => {
   try {
     let address = await DeliveryAddress.find();
-    res.json(address);
+    res.status(200).json(address);
 
   } catch (err) {
     if (err & err.name === 'ValidationError') {
-      return res.json({
+      return res.status(400).json({
         error: 1,
         message: err.message,
         fields: err.errors
@@ -23,11 +25,11 @@ const postDeliveryAddress = async (req, res, next) => {
     let user = req.user;
     let address = new DeliveryAddress({...payload, user: user._id});
     await address.save();
-    return res.json(address);
+    return res.status(201).json(address);
 
   } catch (err) {
     if (err && err.name === 'ValidationError') {
-      return res.json ({
+      return res.status(400).json ({
         error: 1,
         message: err.message,
         fields: err.errors
@@ -42,14 +44,23 @@ const putUpdateDeliveryAddress = async (req, res, next) => {
     let payload = req.body;
     let { id } = req.params;
     let address = await DeliveryAddress.findById(id);
+    let subjectAddress = subject('DeliveryAddress', {...address, user_id: address.user});
+    let policy = policyFor(req.user);
+
+    if (!policy.can('update', subjectAddress)) {
+      return res.status(400).json({
+        error: 1,
+        message: `You're not allowed to modify this resource`
+      });
+    }
 
     address = await DeliveryAddress.findByIdAndUpdate(id, payload, { new: true }
     );
-    return res.json(address);
+    return res.status(200).json(address);
 
   } catch (err) {
     if (err && err.name === 'ValidationError') {
-      return res.json({
+      return res.status(400).json({
         error: 1,
         message: err.message,
         fields: err.errors
@@ -62,7 +73,18 @@ const putUpdateDeliveryAddress = async (req, res, next) => {
 const deleteDeliveryAddressById = async (req, res, next) => {
   try {
     let { id } = req.params;
-    await DeliveryAddress.findByIdAndDelete(id);
+    let address = await DeliveryAddress.findById(id);
+    let subjectAddress = subject('DeliveryAddress', {...address, user_id: address.user});
+    let policy = policyFor(req.user);
+
+    if (!policy.can('delete', subjectAddress)) {
+      return res.status(400).json({
+        error: 1,
+        message: `You're not allowed to modify this resource`
+      });
+    }
+
+    address = await DeliveryAddress.findByIdAndDelete(id);
 
     return res.status(200).json({
       error: 0,
